- using xmYx for names - config cleanup

+++ test/src/main.cpp
@@ -89,7 +89,7 @@ int main(int argc, char** args) {
 			return 0;
 		}
 		if (command == "-sffs") {
-			using T = uint32_t;
+			using T = uint64_t;
 			run_sffs<T>();
 			return 0;
 		}
